requirements.txt
matplotlib==3.6.0
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.2.0
scikit_image==0.19.3
skimage==0.0
tabulate==0.9.0
